+++ app/src/main/java/com/orange/ods/app/ui/components/buttons/icons/ButtonsIcon.kt
@@ -45,7 +45,7 @@ fun ButtonsIcon(customizationState: ButtonIconCustomizationState) {
         Column(
             modifier = Modifier
                 .verticalScroll(rememberScrollState())
-                .padding(vertical = dimensionResource(id = R.dimen.screen_vertical_margin))
+                .padding(vertical = dimensionResource(id = com.orange.ods.R.dimen.screen_vertical_margin))
         ) {
             Row(
                 modifier = Modifier.fillMaxWidth(),
@@ -56,7 +56,7 @@ fun ButtonsIcon(customizationState: ButtonIconCustomizationState) {
                 )
             }
 
-            Spacer(modifier = Modifier.padding(top = dimensionResource(R.dimen.spacing_s)))
+            Spacer(modifier = Modifier.padding(top = dimensionResource(com.orange.ods.R.dimen.spacing_s)))
 
             InvertedBackgroundColumn(horizontalAlignment = Alignment.CenterHorizontally) {
                 IconButton(
@@ -66,7 +66,7 @@ fun ButtonsIcon(customizationState: ButtonIconCustomizationState) {
             }
 
             CodeImplementationColumn(
-                modifier = Modifier.padding(horizontal = dimensionResource(id = R.dimen.screen_horizontal_margin)),
+                modifier = Modifier.padding(horizontal = dimensionResource(id = com.orange.ods.R.dimen.screen_horizontal_margin)),
                 xmlAvailable = true
             ) {
                 FunctionCallCode(
@@ -92,7 +92,12 @@ private fun IconButton(
     val contentDescription = stringResource(id = R.string.component_button_icon_search_desc)
     val onClick = { clickOnElement(context, context.getString(R.string.component_button_icon)) }
 
-    Box(modifier = Modifier.padding(horizontal = dimensionResource(R.dimen.screen_horizontal_margin), vertical = dimensionResource(R.dimen.spacing_m))) {
+    Box(
+        modifier = Modifier.padding(
+            horizontal = dimensionResource(com.orange.ods.R.dimen.screen_horizontal_margin),
+            vertical = dimensionResource(com.orange.ods.R.dimen.spacing_m)
+        )
+    ) {
         UiFramework<OdsIconButtonBinding>(
             compose = {
                 OdsIconButton(