+++ app/src/main/java/com/orange/ods/app/ui/components/buttons/icons/ButtonsIconToggle.kt
@@ -48,7 +48,7 @@ fun ButtonsIconToggle(customizationState: ButtonIconCustomizationState) {
         Column(
             modifier = Modifier
                 .verticalScroll(rememberScrollState())
-                .padding(vertical = dimensionResource(id = R.dimen.screen_vertical_margin))
+                .padding(vertical = dimensionResource(id = com.orange.ods.R.dimen.screen_vertical_margin))
         ) {
             val onCheckedChange: (Boolean) -> Unit = { checked -> buttonCheckedState.value = checked }
 
@@ -63,7 +63,7 @@ fun ButtonsIconToggle(customizationState: ButtonIconCustomizationState) {
                 )
             }
 
-            Spacer(modifier = Modifier.padding(top = dimensionResource(R.dimen.spacing_s)))
+            Spacer(modifier = Modifier.padding(top = dimensionResource(com.orange.ods.R.dimen.spacing_s)))
 
             InvertedBackgroundColumn(horizontalAlignment = Alignment.CenterHorizontally) {
                 IconToggleButton(
@@ -75,7 +75,7 @@ fun ButtonsIconToggle(customizationState: ButtonIconCustomizationState) {
             }
 
             CodeImplementationColumn(
-                modifier = Modifier.padding(horizontal = dimensionResource(id = R.dimen.screen_horizontal_margin)),
+                modifier = Modifier.padding(horizontal = dimensionResource(id = com.orange.ods.R.dimen.screen_horizontal_margin)),
                 xmlAvailable = true
             ) {
                 FunctionCallCode(
@@ -105,7 +105,12 @@ private fun IconToggleButton(
     val checkedIconResId = R.drawable.ic_heart
     val iconContentDescription = stringResource(id = R.string.component_button_icon_toggle_favorite_icon_desc)
 
-    Box(modifier = Modifier.padding(horizontal = dimensionResource(R.dimen.screen_horizontal_margin), vertical = dimensionResource(R.dimen.spacing_m))) {
+    Box(
+        modifier = Modifier.padding(
+            horizontal = dimensionResource(com.orange.ods.R.dimen.screen_horizontal_margin),
+            vertical = dimensionResource(com.orange.ods.R.dimen.spacing_m)
+        )
+    ) {
         UiFramework<OdsIconToogleButtonBinding>(
             compose = {
                 OdsIconToggleButton(