+++ app/src/main/java/com/orange/ods/app/ui/guidelines/spacing/GuidelineSpacing.kt
@@ -13,17 +13,16 @@ package com.orange.ods.app.ui.guidelines.spacing
 import androidx.annotation.DimenRes
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.dimensionResource
-import com.orange.ods.app.R
 
 enum class Spacing(val tokenName: String, @DimenRes val dimenRes: Int) {
 
-    None("spacing - none", R.dimen.spacing_none),
-    ExtraSmall("spacing - xs", R.dimen.spacing_xs),
-    Small("spacing - s", R.dimen.spacing_s),
-    Medium("spacing - m", R.dimen.spacing_m),
-    Large("spacing - l", R.dimen.spacing_l),
-    ExtraLarge("spacing - xl", R.dimen.spacing_xl),
-    ExtraExtraLarge("spacing - 2xl", R.dimen.spacing_2xl);
+    None("spacing - none", com.orange.ods.R.dimen.spacing_none),
+    ExtraSmall("spacing - xs", com.orange.ods.R.dimen.spacing_xs),
+    Small("spacing - s", com.orange.ods.R.dimen.spacing_s),
+    Medium("spacing - m", com.orange.ods.R.dimen.spacing_m),
+    Large("spacing - l", com.orange.ods.R.dimen.spacing_l),
+    ExtraLarge("spacing - xl", com.orange.ods.R.dimen.spacing_xl),
+    ExtraExtraLarge("spacing - 2xl", com.orange.ods.R.dimen.spacing_2xl);
 
     @Composable
     fun getDp() = dimensionResource(id = dimenRes)